Merge groups done. You can now add groups to the game state. After every move, if it is a valid move, it will try to merge groups.
For a single piece:
It removes the group of the single piece and adds it's target coordinate
to the merged group. The merged group contains all the coordinates of
the groups of the neighbours of the target location.

For grouped piece:
Changes the coordinate of the position of the origin coordinate to the
target coordinate in it's group. The merged group contains all the
coordinates of the groups of the neighbours of the target location,
including it's own group.